add drain node button
