+++ assets/src/components/kubernetes/cluster/Nodes.tsx
@@ -4,20 +4,15 @@ import { filesize } from 'filesize'
 import {
   IconFrame,
   SortDescIcon,
-  Switch,
   useSetBreadcrumbs,
 } from '@pluralsh/design-system'
 
-import { useTheme } from 'styled-components'
-
 import {
-  DrainNodeMutationVariables,
   Maybe,
   Node_NodeList as NodeListT,
   Node_Node as NodeT,
   NodesQuery,
   NodesQueryVariables,
-  useDrainNodeMutation,
   useNodesQuery,
 } from '../../../generated/graphql-kubernetes'
 import { ResourceReadyChip, useDefaultColumns } from '../common/utils'
@@ -30,12 +25,9 @@ import {
 } from '../../../routes/kubernetesRoutesConsts'
 import { useCluster } from '../Cluster'
 
-import { Confirm } from '../../utils/Confirm'
-
-import { KubernetesClient } from '../../../helpers/kubernetes.client'
-
 import { getClusterBreadcrumbs } from './Cluster'
 import { UsageBar } from '../../utils/UsageBar.tsx'
+import { DrainNodeModal } from '../common/DrainNodeModal.tsx'
 
 export const getBreadcrumbs = (cluster?: Maybe<KubernetesClusterFragment>) => [
   ...getClusterBreadcrumbs(cluster),
@@ -129,18 +121,7 @@ const colActions = columnHelper.accessor(() => null, {
   id: 'actions',
   header: '',
   cell: function Cell({ row: { original } }) {
-    const theme = useTheme()
-    const cluster = useCluster()
     const [open, setOpen] = useState(false)
-    const [ignoreAllDaemonSets, setIgnoreAllDaemonSets] = useState(true)
-    const [mutation, { loading, error }] = useDrainNodeMutation({
-      client: KubernetesClient(cluster?.id ?? ''),
-      variables: {
-        name: original.objectMeta.name ?? '',
-        input: { ignoreAllDaemonSets },
-      } as DrainNodeMutationVariables,
-      onCompleted: () => setOpen(false),
-    })
 
     return (
       <>
@@ -153,28 +134,11 @@ const colActions = columnHelper.accessor(() => null, {
             setOpen(true)
           }}
         />
-        {open && (
-          <Confirm
-            close={() => setOpen(false)}
-            destructive
-            label="Drain node"
-            loading={loading}
-            error={error}
-            open={open}
-            submit={() => mutation()}
-            title="Drain node"
-            text={`Are you sure you want to drain ${original?.objectMeta.name} node? Node will be cordoned first. Please note that it may take a while to complete.`}
-            extraContent={
-              <Switch
-                checked={ignoreAllDaemonSets}
-                onChange={setIgnoreAllDaemonSets}
-                css={{ paddingTop: theme.spacing.medium }}
-              >
-                Ignore all daemon sets
-              </Switch>
-            }
-          />
-        )}
+        <DrainNodeModal
+          name={original.objectMeta.name ?? ''}
+          open={open}
+          setOpen={setOpen}
+        />
       </>
     )
   },
